requirements.txt
openai==<1.55.0>
pydantic==<1.10.13>
python-dotenv==<1.0.1>